Remove unused code, fix circular deps
