+++ packages/react-native-sortables/src/components/shared/SortableContainer.tsx
@@ -1,3 +1,4 @@
+import { memo, useSyncExternalStore } from 'react';
 import type { StyleProp, ViewStyle } from 'react-native';
 import Animated, {
   LinearTransition,
@@ -8,14 +9,15 @@ import Animated, {
 import { EMPTY_OBJECT, IS_WEB } from '../../constants';
 import { DebugOutlet } from '../../debug';
 import type { AnimatedStyleProp } from '../../integrations/reanimated';
-import { ItemsOutlet, useCommonValuesContext } from '../../providers';
+import { useCommonValuesContext, useItemsContext } from '../../providers';
 import type {
   DimensionsAnimation,
   DropIndicatorSettings,
   Overflow
 } from '../../types';
-import type { DraggableViewProps } from '..';
 import AnimatedOnLayoutView from './AnimatedOnLayoutView';
+import type { DraggableViewProps } from './DraggableView';
+import { DraggableView } from './DraggableView';
 import DropIndicator from './DropIndicator';
 
 export type SortableContainerProps = DropIndicatorSettings &
@@ -124,3 +126,28 @@ export default function SortableContainer({
     </Animated.View>
   );
 }
+
+type ItemsOutletProps = Pick<
+  DraggableViewProps,
+  'itemEntering' | 'itemExiting'
+> & {
+  itemStyle?: AnimatedStyleProp;
+};
+
+const ItemsOutlet = memo(function ItemsOutlet({
+  itemStyle,
+  ...rest
+}: ItemsOutletProps) {
+  const { getKeys, subscribeKeys } = useItemsContext();
+
+  // Re-render the list of cells only if the keys array changes
+  const keys = useSyncExternalStore(
+    subscribeKeys,
+    getKeys,
+    getKeys // SSR fallback
+  );
+
+  return keys.map(key => (
+    <DraggableView {...rest} itemKey={key} key={key} style={itemStyle} />
+  ));
+});
