--- packages/react-native-sortables/src/providers/shared/ItemsProvider/components/index.ts
@@ -1,2 +1,0 @@
-export { default as ItemOutlet } from './ItemOutlet';
-export { default as ItemsOutlet } from './ItemsOutlet';
